fix: remove obsolete component and add class.
